refactor(registry): extract blob and token adapters out of store.py
Registry RFC phase A. MemoryBlobStore / FilesystemBlobStore / S3BlobStore
move to blobs.py; the TokenStore family (TokenInfo, scope constants,
PersistentTokenStore + SQLite/Postgres variants) moves to tokens.py with
its api_tokens schema bootstrap. store.py re-exports them so importers
stay untouched until phase C. Zero behavior change.
